Hydrometer Calc Rearranged some of the code for readability and easier management. Minor update to the interface.
Bitterness Calc
Significant rewrite to much of the code.
Modified interface to allow dynamic additions of hop entry fields.
Max number of hop entry fields is controlled by a constant at
the top of the file: $MAX_HOPS.
Moved all constants to the top.
Moved all JavaScript to bitterness.js since there's a lot more of
it now - the hop entry field stuff is all JS.
Modified the interface to make Garetz calculations optional.
Modifications to $pelletFactor. This is a constant at the top
right now but will soon be set in prefs.
Moved all IBU formulas into a separate library, bitterness.lib.php.
This will allow all IBU calculations to be handled in one file
instead of duplicating the code. The next step is to convert the
recipe/blog IBU code to start using it.

Updated all relevant files to link to the two calculators which
were modified/added.