Fix Recipe Display Issues
Storage of recipe data in DB will be converted to a single row for fermentables, adjuncts, hops, water, and yeast (currently, most are multiple rows housing a single instance of an ingredient). Previous programming displayed the single row data - this needed to be "rolled back" due to display issues. Once the Admin side code is updated to use the new schema, front-end display will be updated as well.